fix: remove duplicates when searching Closes #75
