+++ src/main/frontend/src/launcher-quarkus/pickers/extensions-picker.tsx
@@ -3,6 +3,7 @@ import { Button, FormGroup, TextInput, Tooltip } from "@patternfly/react-core";
 import { CheckSquareIcon, ClipboardCheckIcon, ClipboardIcon, OutlinedSquareIcon, SearchIcon, TrashAltIcon } from "@patternfly/react-icons";
 import copy from 'copy-to-clipboard';
 import React, { useState } from "react";
+import { processEntries } from './extensions-picker-helpers';
 import './extensions-picker.scss';
 
 export interface ExtensionEntry {
@@ -99,42 +100,6 @@ function Extension(props: ExtensionProps) {
   )
 }
 
-export const filterFunction = (filter: string) => (d: ExtensionEntry) => {
-  const filterLowerCase = filter.trim().toLowerCase();
-  if (!filterLowerCase) {
-    return true;
-  }
-  const shortName = d.shortName ? d.shortName.toLowerCase() : '';
-  if (filterLowerCase === shortName) {
-    return true;
-  }
-  return d.name.toLowerCase().includes(filterLowerCase)
-    || d.labels.filter(l => l.startsWith(filterLowerCase)).length > 0
-    || (d.category && d.category.toLowerCase().startsWith(filterLowerCase))
-    || shortName.startsWith(filterLowerCase);
-}
-
-export const sortFunction = (filter: string) => (a: ExtensionEntry, b: ExtensionEntry) => {
-  const filterLowerCase = filter.trim().toLowerCase();
-  if (!filterLowerCase) {
-    return a.order > b.order ? 1 : -1;
-  }
-  const startWithAShortName = !!a.shortName && a.shortName.toLowerCase().startsWith(filterLowerCase);
-  const startWithBShortName = !!b.shortName && b.shortName.toLowerCase().startsWith(filterLowerCase);
-  if (startWithAShortName !== startWithBShortName) {
-    return startWithAShortName ? -1 : 1;
-  }
-  const startWithOneOfALabel = a.labels.filter(l => l.startsWith(filterLowerCase)).length > 0;
-  const startWithOneOfBLabel = b.labels.filter(l => l.startsWith(filterLowerCase)).length > 0;
-  if (startWithOneOfALabel !== startWithOneOfBLabel) {
-    return startWithOneOfALabel ? -1 : 1;
-  }
-  if (a.name.toLowerCase().startsWith(filterLowerCase) !== b.name.toLowerCase().startsWith(filterLowerCase)) {
-    return a.name.toLowerCase().startsWith(filterLowerCase) ? -1 : 1;
-  }
-  return a.order > b.order ? 1 : -1;
-}
-
 export const ExtensionsPicker: Picker<ExtensionsPickerProps, ExtensionsPickerValue> = {
   checkCompletion: (value: ExtensionsPickerValue) => !!value.extensions && value.extensions.length > 0,
   Element: (props: ExtensionsPickerProps) => {
@@ -164,8 +129,7 @@ export const ExtensionsPicker: Picker<ExtensionsPickerProps, ExtensionsPickerVal
       setHasSearched(true);
       setFilter(f);
     }
-
-    const result = props.entries.filter(filterFunction(filter));
+    const result = processEntries(filter, props.entries);
     const categories = new Set(props.entries.map(i => i.category));
     let currentCat: string | undefined = undefined;
     return (
@@ -211,7 +175,7 @@ export const ExtensionsPicker: Picker<ExtensionsPickerProps, ExtensionsPickerVal
             </div>
           )}
           <div className="list-container">
-            {result.sort(sortFunction(filter)).map((ex, i) => {
+            {result.map((ex, i) => {
               const ext = (
                 <Extension
                   selected={entrySet.has(ex.id)}
